+++ crates/serialization/src/arrow/trade.rs
@@ -239,7 +239,7 @@ mod tests {
     use std::sync::Arc;
 
     use arrow::{
-        array::{Array, FixedSizeBinaryArray, UInt8Array, UInt64Array},
+        array::{Array, UInt8Array, UInt64Array},
         record_batch::RecordBatch,
     };
     use nautilus_model::types::{
@@ -248,7 +248,7 @@ mod tests {
     use rstest::rstest;
 
     use super::*;
-    use crate::arrow::{get_raw_price, get_raw_quantity};
+    use crate::arrow::{fixed_size_binary, get_raw_price, get_raw_quantity};
 
     #[rstest]
     fn test_get_schema() {
@@ -382,10 +382,9 @@ mod tests {
 
         let raw_price1 = (100.00 * FIXED_SCALAR) as PriceRaw;
         let raw_price2 = (101.00 * FIXED_SCALAR) as PriceRaw;
-        let price =
-            FixedSizeBinaryArray::from(vec![&raw_price1.to_le_bytes(), &raw_price2.to_le_bytes()]);
+        let price = fixed_size_binary(vec![&raw_price1.to_le_bytes(), &raw_price2.to_le_bytes()]);
 
-        let size = FixedSizeBinaryArray::from(vec![
+        let size = fixed_size_binary(vec![
             &((1000.0 * FIXED_SCALAR) as QuantityRaw).to_le_bytes(),
             &((900.0 * FIXED_SCALAR) as QuantityRaw).to_le_bytes(),
         ]);
@@ -421,8 +420,8 @@ mod tests {
         let metadata = TradeTick::get_metadata(&instrument_id, 2, 0);
 
         let raw_price = (100.00 * FIXED_SCALAR) as PriceRaw;
-        let price = FixedSizeBinaryArray::from(vec![&raw_price.to_le_bytes()]);
-        let size = FixedSizeBinaryArray::from(vec![
+        let price = fixed_size_binary(vec![&raw_price.to_le_bytes()]);
+        let size = fixed_size_binary(vec![
             &((1000.0 * FIXED_SCALAR) as QuantityRaw).to_le_bytes(),
         ]);
         let aggressor_side = UInt8Array::from(vec![0]);
@@ -470,8 +469,8 @@ mod tests {
         let metadata = TradeTick::get_metadata(&instrument_id, 2, 0);
 
         let invalid_price: PriceRaw = PriceRaw::MAX - 1000;
-        let price = FixedSizeBinaryArray::from(vec![&invalid_price.to_le_bytes()]);
-        let size = FixedSizeBinaryArray::from(vec![
+        let price = fixed_size_binary(vec![&invalid_price.to_le_bytes()]);
+        let size = fixed_size_binary(vec![
             &((1000.0 * FIXED_SCALAR) as QuantityRaw).to_le_bytes(),
         ]);
         let aggressor_side = UInt8Array::from(vec![0]);
@@ -511,10 +510,10 @@ mod tests {
         let metadata = TradeTick::get_metadata(&instrument_id, 2, FIXED_PRECISION);
 
         let raw_price = (100.00 * FIXED_SCALAR) as PriceRaw;
-        let price = FixedSizeBinaryArray::from(vec![&raw_price.to_le_bytes()]);
+        let price = fixed_size_binary(vec![&raw_price.to_le_bytes()]);
 
         let invalid_size = QUANTITY_RAW_MAX + 1;
-        let size = FixedSizeBinaryArray::from(vec![&invalid_size.to_le_bytes()]);
+        let size = fixed_size_binary(vec![&invalid_size.to_le_bytes()]);
         let aggressor_side = UInt8Array::from(vec![0]);
         let trade_id = StringArray::from(vec!["1"]);
         let ts_event = UInt64Array::from(vec![1]);
@@ -548,8 +547,8 @@ mod tests {
         let metadata = TradeTick::get_metadata(&instrument_id, 2, 0);
 
         let raw_price = (100.00 * FIXED_SCALAR) as PriceRaw;
-        let price = FixedSizeBinaryArray::from(vec![&raw_price.to_le_bytes()]);
-        let size = FixedSizeBinaryArray::from(vec![
+        let price = fixed_size_binary(vec![&raw_price.to_le_bytes()]);
+        let size = fixed_size_binary(vec![
             &((1000.0 * FIXED_SCALAR) as QuantityRaw).to_le_bytes(),
         ]);
 
@@ -587,8 +586,8 @@ mod tests {
         metadata.remove(KEY_INSTRUMENT_ID);
 
         let raw_price = (100.00 * FIXED_SCALAR) as PriceRaw;
-        let price = FixedSizeBinaryArray::from(vec![&raw_price.to_le_bytes()]);
-        let size = FixedSizeBinaryArray::from(vec![
+        let price = fixed_size_binary(vec![&raw_price.to_le_bytes()]);
+        let size = fixed_size_binary(vec![
             &((1000.0 * FIXED_SCALAR) as QuantityRaw).to_le_bytes(),
         ]);
         let aggressor_side = UInt8Array::from(vec![0]);
@@ -625,8 +624,8 @@ mod tests {
         metadata.remove(KEY_PRICE_PRECISION);
 
         let raw_price = (100.00 * FIXED_SCALAR) as PriceRaw;
-        let price = FixedSizeBinaryArray::from(vec![&raw_price.to_le_bytes()]);
-        let size = FixedSizeBinaryArray::from(vec![
+        let price = fixed_size_binary(vec![&raw_price.to_le_bytes()]);
+        let size = fixed_size_binary(vec![
             &((1000.0 * FIXED_SCALAR) as QuantityRaw).to_le_bytes(),
         ]);
         let aggressor_side = UInt8Array::from(vec![0]);